pyproject.toml
[project]
name = "autism-crawler"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "praw>=7.7.0",
    "pgvector>=0.3.0",
    "openai>=1.0.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
admin = [
    "django>=5.0.0",
    "psycopg2-binary>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
